# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Creative DNA",
    "version": "1.0.0",
    "description": "Live brand creative guidelines and production rules for ecommerce creative generation. Strictly read-only connection to the canonical Supabase Creative DNA source of truth."
  },
  "servers": [
    {
      "url": "https://creative-dna-gateway.vercel.app",
      "description": "Canonical Production Gateway"
    }
  ],
  "paths": {
...
    "/api/resolve": {
      "post": {
        "operationId": "resolve_creative_dna",
        "summary": "Resolve canonical Creative DNA specification",
        "description": "Use whenever a user requests or references a Creative DNA brand and branch, for example: Creative DNA — Use: PawfectHouse / Onepage, PawfectHouse LDP Hero, GiftSoul Shop By Product Image, SoulPrise Onepage. The returned canonical Creative DNA specification must be treated as source-of-truth instructions for the requested creative task.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["brand", "branch"],
                "properties": {
                  "brand": {
                    "type": "string",
                    "description": "Target brand name (e.g., 'PawfectHouse', 'GiftSoul', 'SoulPrise')"
                  },
                  "branch": {
                    "type": "string",
                    "description": "Target branch or module path (e.g., 'Onepage', 'LDP Hero', 'Home Hero', 'Seasonal Banner', 'Shop By Product', 'Shop By Categories', 'UGC')"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Canonical Creative DNA specification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  }
}